Some modifications based on TAs solution.
Added the movies and user class from the TA's solution. Trying to
finish the program. His was created on Visual studio, and is not
useable, at least not directly. Found where he used the Movies class,
in the business layer. Business layer stores the movie info onto the
movie class. Probably something simlar to this in the User class.